+++ main.go
@@ -26,7 +26,7 @@ import (
 )
 
 // Global variables
-var CLIENTVERSION = "v1.2.4"
+var CLIENTVERSION = "v1.3.0"
 var SPECVERSION = 16
 var clientSettings = ClientSettings{
 	CacheDirectory:             "cache/", // Default cache directory