+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+- [2020-07-18] Fixed incorrect reported disk space to server for edge cases by [@lflare].
 
 ### Security